+++ src/components/appbar/index.js
@@ -67,6 +67,7 @@ export default function SimpleTabs() {
             ))}
         </Tabs>
       </AppBar>
+
     </div>
   );
 }
