+++ src/home/index.js
@@ -1,105 +1,131 @@
 import React from 'react';
-import { makeStyles, Button } from '@material-ui/core'
-import TopCollections from './top';
-import NFT from '@/components/NFT';
+import { Button, Skeleton, Space, List, Table } from 'antd';
+import './index.less';
 import Images from '@/constant'
 
-export default function Home(props) {
-    console.log('home')
-    const classes = useStyle();
-    return (
-        <div>
-            <div className={classes.itemCenter}>
-                <div className={classes.left}>
-                    <div className={classes.title}>Trade to Earn,<br />Keep for More</div>
-                    <div className={classes.subTitle}>Buy or sell NFTs to earn $NONO<br />
-                        50% trading fee income will be used for $NONO buyback<br />
-                        List your NFT or explore the market to get started</div>
-                    <div className={classes.box}>
-                        <Button  component="span" textAllCaps={false} disableRipple={true} classes={{ root: classes.buttonAsset }}>List an NFT</Button>
-                        <Button  component="span" disableRipple={true} >Explore NFTs</Button>
-                    </div>
-                </div>
-                <NFT />
-            </div>
-            <TopCollections />
-            <div className={classes.itemCenter}>
-                <div className={classes.left}>
-                    <div className={classes.title}>Ready to Get Your Rewards?</div>
-                    <div className={classes.subTitle}>
-                    Our Users Should Be Our Holders,<br />
-                    And They All Deserves Our Profit Shares.
-                    </div>
-                    <div className={classes.box}>
-                        <Button   textAllCaps={false} disableRipple={true} classes={{ root: classes.buttonAsset }}>Claim Your Rewards</Button>
-                        <Button disableRipple={true}>Learn More</Button>
-                    </div>
-                </div>
-                <img className={classes.rewardImg} src={Images.rewardBanner} />
+const data = [
+    {
+      key: '1',
+      name: 'John Brown',
+      age: 32,
+      address: 'New York No. 1 Lake Park',
+      tags: ['nice', 'developer'],
+    },
+  ];
 
+  const listData = [
+    {
+      title: '1. What is $FAME?',
+      description: '$FAME is the inaugural token launched based on the FAME Protocol.',
+    },
+    {
+        title: '1. What is $FAME?',
+        description: '$FAME is the inaugural token launched based on the FAME Protocol.',
+      },
+      {
+        title: '1. What is $FAME?',
+        description: '$FAME is the inaugural token launched based on the FAME Protocol.',
+      },
+      {
+        title: '1. What is $FAME?',
+        description: '$FAME is the inaugural token launched based on the FAME Protocol.',
+      },
+  ];
+export default function Explore(props) {
+    console.log('Explore')
+
+    
+    const columns = [
+        {
+          title: 'Token',
+          dataIndex: 'name',
+          key: 'name',
+          render: (text) => {
+                return <div style={{width: '100%', height: 100, display: 'flex', flexDirection: 'row'}}>
+                    <div>
+                        <img src={Images.logo} alt='' style={{ width: 60, height: 60 }} />
+                    </div>
+                    <div style={{display: 'flex', flexDirection: 'column'}}>
+                    <div>pizza</div>
+                    <div>pizza2</div>
+                    </div>
+                </div>
+            },
+        },
+        {
+          title: 'Progress',
+          dataIndex: 'age',
+          key: 'age',
+        },
+        {
+          title: 'Status',
+          dataIndex: 'address',
+          key: 'address',
+        },
+        {
+            title: 'Created',
+            dataIndex: 'address',
+            key: 'address',
+          },
+        {
+          title: 'Action',
+          key: 'action',
+          render: (_, record) => (
+            <div style={{ display: 'flex', flexDirection: 'column'}}>
+              <Button type='primary' onClick={() => {}} style={{marginBottom: 10}}>Swep</Button>
+
+              <Button type='primary' onClick={() => {}}>Detail</Button>
             </div>
+          ),
+        },
+      ];
+    return (
+        <div className={'main'}>
+            <div className='title1'>FAME Protocol</div>
+            <div>A Fair Launch Protocol for meme coins, base on Solana.</div>
+            <div>
+                <Button type='primary' onClick={() => { }}>Deploy a new coin</Button>
+                <Button onClick={() => { }}>Deploy a new coin</Button>
+            </div>
+            <div>Pool Created</div>
+            <div style={{ width: '100%' }}>
+                <Table
+                    columns={columns}
+                    dataSource={data}
+                    pagination={false}
+                    footer={(currentPageData => {
+                        return <a onClick={() => { }}
+                            style={{ width: '100%', display: 'flex', flexDirection: 'row', justifyContent: 'center' }}>
+                            加载更多
+                        </a>
+                    })}
+
+                />
+            </div>
+
+            <div style={{ width: '100%' }}>
+                <List
+                    itemLayout="horizontal"
+                    dataSource={listData}
+                    renderItem={item => (
+                        <List.Item>
+                            <List.Item.Meta
+                                title={<a href="https://ant.design">{item.title}</a>}
+                                description={item.description}
+                            />
+                        </List.Item>
+                    )}
+                />
+            </div>
+
+            <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
+                <img src={Images.logo} alt='' style={{ width: 60, height: 60, }} />
+                <img src={Images.logo} alt='' style={{ width: 60, height: 60 }} />
+            </div>
+            <footer>
+                <img src={Images.logo} alt='' style={{ width: 60, height: 60 }} />
+            </footer>
+            <footer>© 2024 by FAME. All rights reserved!</footer>
         </div>
     )
-};
-
-const useStyle = makeStyles((theme) => ({
-    itemCenter: {
-        background: '#fff',
-        borderRadius: '20px',
-        display: 'flex',
-        justifyContent: 'space-between',
-        padding: '69px 50px 69px 100px',
-        alignItems: 'center',
-        flexWrap: 'wrap',
-        [theme.breakpoints.between('sm', 'md')]: {
-            padding: theme.custom.palette.mdspacing,
-        },
-    },
-    box: {
-        flex: 1,
-        [theme.breakpoints.between('sm', 'md')]: {
-            display: 'flex',
-            flexDirection: 'column',
-            '& button': {
-                fontSize: '14px',
-                marginBottom: theme.custom.palette.mdspacing,
-
-            }
-        },
-    },
-    title: {
-        fontFamily: 'Archivo Black',
-        fontWeight: 400,
-        fontSize: '36px',
-        lineHeight: '54px',
-        color: '#000000',
-        paddingBottom: '50px',
-        [theme.breakpoints.between('sm', 'md')]: {
-            paddingBottom: theme.custom.palette.mdspacing,
-            fontSize: '24px',
-
-        },
-    },
-    subTitle: {
-        fontFamily: 'Barlow',
-        fontWeight: 600,
-        fontSize: '18px',
-        lineHeight: '27px',
-        color: '#333333',
-        paddingBottom: '50px',
-        [theme.breakpoints.between('sm', 'md')]: {
-            paddingBottom: theme.custom.palette.mdspacing,
-        },
-    },
-    buttonAsset: {
-        background: '#000',
-        color: '#fff',
-        marginRight: '30px',
-        '&:hover': {
-            background: '#000',
-        }
-    },
-    rewardImg: {
-       marginRight: '237px', 
-    }
-}))
+}
