+++ src/index.js
@@ -7,7 +7,9 @@ import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
+
 import "./index.css";
+import 'antd/dist/antd.css';
 import store from "./redux/store";
 import reportWebVitals from "./reportWebVitals";
